feat: Replace ValueError with UnregisteredDagError in scheduling logic and add unit tests for SchedulingHelper
